Toevoegen eigenschap enkelfout + niet inserten als regel al bestaat

+++ whine_DB_SDK.py
@@ -1,51 +1,54 @@
 import sqlite3
 from whine_classes import WhineBottle
 
+# gebruiken we overal dus global
+conn = sqlite3.connect('whine_inventory.db')
+c = conn.cursor()
 
-def create_table():
-    conn = sqlite3.connect('whine_inventory.db')
-    c = conn.cursor()
+def create_table(drop):
+    if drop == True:
+        c.execute('DROP TABLE IF EXISTS whine_bottles')
+        c.execute('DROP TABLE IF EXISTS bottle_properties')
     c.execute('CREATE TABLE IF NOT EXISTS whine_bottles (UID TEXT PRIMARY KEY, name TEXT, main_grape TEXT, year TEXT, date_in_fridge DATE)')
-    c.execute('CREATE TABLE IF NOT EXISTS bottle_properties (UID TEXT PRIMARY KEY, property TEXT, value TEXT)')
+    c.execute('CREATE TABLE IF NOT EXISTS bottle_properties (property_id integer PRIMARY KEY AUTOINCREMENT, UID TEXT,  property TEXT, value TEXT)')
 	
 def recreate_table():
-    conn = sqlite3.connect('whine_inventory.db')
-    c = conn.cursor()
-    c.execute('DROP TABLE IF EXISTS whine_bottles')
-    c.execute('DROP TABLE IF EXISTS bottle_properties')
-    create_table()
+    create_table(True)
 
 def add_whine(UID, name, main_grape, year, properties, date_in_fridge):
-    conn = sqlite3.connect('whine_inventory.db')
-    c = conn.cursor()
-    c.execute('INSERT INTO whine_bottles (UID, name, main_grape, year, date_in_fridge) VALUES (?, ?, ?, ?, ?)', (UID, name, main_grape, year, date_in_fridge))
-    conn.commit()
-    #Add bottle properties
-    add_whine_properties(UID, properties)
-    message = print('Succesfully inserted new bottle!')
-    return message
+    c.execute("SELECT UID from bottle_properties where UID = '" + UID + "'")
+    data = c.fetchone()
+    if data:
+        # regel is al gevonden, dus niet opnieuw inserten
+        message = print("Fles met tag " + UID + " bestaat al, dus inserten gaat niet door.")
+    else:
+        c.execute('INSERT INTO whine_bottles (UID, name, main_grape, year, date_in_fridge) VALUES (?, ?, ?, ?, ?)', (UID, name, main_grape, year, date_in_fridge))
+        conn.commit()
+        #Add bottle properties
+        #add_whine_properties(UID, properties)
+        message = print('Succesfully inserted new bottle!')
+        return message
 
-def add_whine_properties(UID, properties):
-    conn = sqlite3.connect('whine_inventory.db')
-    c = conn.cursor()
-    if properties:
-        for key, val in properties.items(): #<--- dit krijg ik nog niet werkend. Ik wil de value van de dict wil ik gabruiken in de query. Als ik properties[val] doe krijg ik een key error.
-            c.execute('INSERT INTO bottle_properties (UID, property, value) VALUES (?, ?, ?)', (UID, properties[key], properties[key]))
+def add_whine_property(UID, property, value):
+    if property is not None and value is not None:
+        c.execute("SELECT property from bottle_properties where property = '"+property+"'")
+        data = c.fetchone()
+        if data:
+            # regel is al gevonden, dus niet opnieuw inserten
+            message = print("Eigenschap " + property + " bestaat al voor deze fles, dus inserten gaat niet door.")
+        else:
+            c.execute('INSERT INTO bottle_properties (UID, property, value) VALUES (?, ?, ?)', (UID, property, value))
             conn.commit()
-            message = print("Succesfully inserted bottle "+UID+"\'s properties")
+            message = print("Succesfully inserted bottle "+UID+"\'s property "+property)
             return message
 
 def fetch_all_results():
-    conn = sqlite3.connect('whine_inventory.db')
-    c = conn.cursor()
     c.execute('''SELECT UID, name, main_grape, year, date_in_fridge FROM whine_bottles ORDER BY date_in_fridge DESC''')
     data = c.fetchall()
     return data
 	
 def fetch_bottle(UID):
     print("Fetching bottle with UID="+UID)
-    conn = sqlite3.connect('whine_inventory.db')
-    c = conn.cursor()
     c.execute("SELECT name, main_grape, year, date_in_fridge FROM whine_bottles WHERE UID='"+UID+"'")
     data = c.fetchone()
     if data:
@@ -68,8 +71,6 @@ def fetch_bottle(UID):
 
 def fetch_bottle_properties(UID):
     print("Fetching bottle with UID="+UID)
-    conn = sqlite3.connect('whine_inventory.db')
-    c = conn.cursor()
     c.execute("SELECT UID, property, value FROM bottle_properties WHERE UID='"+UID+"'")
     data = c.fetchone()
     if data:
@@ -90,15 +91,11 @@ def fetch_bottle_properties(UID):
 
 
 def delete_selected(UID):
-    conn = sqlite3.connect('whine_inventory.db')
-    c = conn.cursor()
     c.execute("DELETE FROM whine_bottles WHERE UID='"+UID+"'")
     conn.commit()
     return print('Succes deleted bottle with UID' + UID) 
 
 def clear_results():
-    conn = sqlite3.connect('whine_inventory.db')
-    c = conn.cursor()
     c.execute('''DELETE FROM whine_bottles ''')
     conn.commit()
     return print('Succes truncated table \'whine_bottles\'!')
